feat: adjust menu
