+++ containers/Ai/views/credentials/index.vue
@@ -11,16 +11,12 @@
 
 <script>
 import AccessKeyList from './components/AccessKeyList'
-import OpenIdList from './components/OpenIdList'
-import EncryptKeyList from './components/EncryptKeyList'
 import ContainerSecretList from '../container-secret/components/List'
 
 export default {
   name: 'CredentialIndex',
   components: {
     AccessKeyList,
-    OpenIdList,
-    EncryptKeyList,
     ContainerSecretList,
   },
   data () {
@@ -30,14 +26,6 @@ export default {
         {
           key: 'AccessKeyList',
           label: this.$t('scope.text_4'),
-        },
-        {
-          key: 'EncryptKeyList',
-          label: this.$t('scope.title.encrypt_key'),
-        },
-        {
-          key: 'OpenIdList',
-          label: 'OIDC/OAuth2',
         },
         {
           key: 'ContainerSecretList',
