+++ containers/Monitor/router/index.js
@@ -1,5 +1,5 @@
 import Layout from '@/layouts/RouterView'
-import { setupKeys } from '@/utils/auth'
+import { setupKeys, hasServices } from '@/utils/auth'
 import i18n from '@/locales'
 import store from '@/store'
 import { isScopedPolicyMenuHidden } from '@/utils/scopedPolicy'
@@ -16,6 +16,13 @@ const Explorer = () => import(/* webpackChunkName: "monitor" */ /* webpackPrefet
 const Dashboard = () => import(/* webpackChunkName: "monitor" */ /* webpackPrefetch: true */ '@Monitor/views/dashboard')
 const MonitorDashboardChartCreate = () => import(/* webpackChunkName: "monitor" */ /* webpackPrefetch: true */ '@Monitor/views/dashboard/create')
 const AlertRecordShieldsIndex = () => import(/* webpackChunkName: "monitor" */ /* webpackPrefetch: true */ '@Monitor/views/alertrecordshields')
+
+const Notification = () => import(/* webpackChunkName: "iam" */ /* webpackPrefetch: true */ '@IAM/views/notification')
+const NotifyTopic = () => import(/* webpackChunkName: "iam" */ /* webpackPrefetch: true */ '@IAM/views/notify-topic')
+const Robots = () => import(/* webpackChunkName: "iam" */ /* webpackPrefetch: true */ '@IAM/views/robots')
+const NotifyConfigs = () => import(/* webpackChunkName: "iam" */ /* webpackPrefetch: true */ '@IAM/views/notifyconfig')
+const NotifyConfigCreate = () => import(/* webpackChunkName: "iam" */ /* webpackPrefetch: true */ '@IAM/views/notifyconfig/create')
+const Contact = () => import(/* webpackChunkName: "iam" */ /* webpackPrefetch: true */ '@IAM/views/contact')
 
 export default {
   index: 65,
@@ -286,6 +293,131 @@ export default {
               name: 'AlertRecordShieldsIndex',
               path: '',
               component: AlertRecordShieldsIndex,
+            },
+          ],
+        },
+      ],
+    },
+    {
+      meta: {
+        label: i18n.t('system.text_564'),
+      },
+      submenus: [
+        {
+          path: '/notification',
+          meta: {
+            label: i18n.t('system.text_16'),
+            permission: 'notifications_list',
+            hidden: () => {
+              if (isScopedPolicyMenuHidden('sub_hidden_menus.notification')) {
+                return true
+              }
+              return !hasServices('notify')
+            },
+            t: 'dictionary.webconsole',
+          },
+          component: Layout,
+          children: [
+            {
+              name: 'Notification',
+              path: '',
+              component: Notification,
+            },
+          ],
+        },
+        {
+          path: '/notify-topic',
+          meta: {
+            label: i18n.t('dictionary.notify-topic'),
+            permission: 'topics_list',
+            hidden: () => {
+              if (isScopedPolicyMenuHidden('sub_hidden_menus.notify_topic')) {
+                return true
+              }
+              if (!(store.getters.isAdminMode || store.getters.isDomainMode)) {
+                return true
+              }
+              return !hasServices('notify')
+            },
+            t: 'dictionary.notify-topic',
+          },
+          component: Layout,
+          children: [
+            {
+              name: 'NotifyTopic',
+              path: '',
+              component: NotifyTopic,
+            },
+          ],
+        },
+        {
+          path: '/notifyconfig',
+          meta: {
+            label: i18n.t('system.text_19'),
+            permission: 'notifyconfigs_list',
+            hidden: () => {
+              if (isScopedPolicyMenuHidden('sub_hidden_menus.notifyconfig')) {
+                return true
+              }
+              if (!(store.getters.isAdminMode || store.getters.isDomainMode)) {
+                return true
+              }
+              return !hasServices('notify')
+            },
+          },
+          component: Layout,
+          children: [
+            {
+              name: 'NotifyConfig',
+              path: '',
+              component: NotifyConfigs,
+            },
+            {
+              name: 'NotifyConfigCreate',
+              path: 'create',
+              component: NotifyConfigCreate,
+            },
+          ],
+        },
+        {
+          path: '/contact',
+          meta: {
+            label: i18n.t('common_27'),
+            permission: 'contacts_list',
+            hidden: () => {
+              if (isScopedPolicyMenuHidden('sub_hidden_menus.contact')) {
+                return true
+              }
+              return !store.getters.isAdminMode && !store.getters.isDomainMode
+            },
+          },
+          component: Layout,
+          children: [
+            {
+              name: 'Contact',
+              path: '',
+              component: Contact,
+            },
+          ],
+        },
+        {
+          path: '/robot',
+          meta: {
+            label: i18n.t('system.robot_manage'),
+            permission: 'robots_list',
+            hidden: () => {
+              if (isScopedPolicyMenuHidden('sub_hidden_menus.robot')) {
+                return true
+              }
+              return false
+            },
+          },
+          component: Layout,
+          children: [
+            {
+              name: 'Robots',
+              path: '',
+              component: Robots,
             },
           ],
         },
